+++ lnd_services.go
@@ -1,393 +1,14 @@
 package taprootassets
 
 import (
-	"bytes"
 	"context"
-	"fmt"
-	"sync"
-	"time"
 
-	"github.com/btcsuite/btcd/btcec/v2"
-	"github.com/btcsuite/btcd/chaincfg/chainhash"
-	"github.com/btcsuite/btcd/wire"
 	"github.com/lightninglabs/lndclient"
-	"github.com/lightninglabs/neutrino/cache/lru"
-	"github.com/lightninglabs/taproot-assets/asset"
-	"github.com/lightninglabs/taproot-assets/fn"
-	"github.com/lightninglabs/taproot-assets/proof"
 	"github.com/lightninglabs/taproot-assets/rfq"
 	"github.com/lightninglabs/taproot-assets/tapchannel"
-	"github.com/lightninglabs/taproot-assets/tapdb"
-	"github.com/lightninglabs/taproot-assets/tapgarden"
-	"github.com/lightningnetwork/lnd/chainntnfs"
-	"github.com/lightningnetwork/lnd/funding"
 	"github.com/lightningnetwork/lnd/lnrpc/routerrpc"
-	"github.com/lightningnetwork/lnd/lnwallet/chainfee"
 	"github.com/lightningnetwork/lnd/lnwire"
-	"github.com/lightningnetwork/lnd/protofsm"
-	"github.com/lightningnetwork/lnd/routing/route"
 )
-
-const (
-	// maxNumBlocksInCache is the maximum number of blocks we'll cache
-	// timestamps for. With 400k blocks we should only take up approximately
-	// 3200kB of memory (4 bytes for the block height and 4 bytes for the
-	// timestamp, not including any map/cache overhead).
-	maxNumBlocksInCache = 400_000
-
-	// DefaultTimeout is the default timeout we use for RPC and database
-	// operations.
-	DefaultTimeout = 30 * time.Second
-)
-
-// cacheableTimestamp is a wrapper around an uint32 that can be used as a value
-// in an LRU cache.
-type cacheableTimestamp uint32
-
-// Size returns the size of the cacheable timestamp. Since we scale the cache by
-// the number of items and not the total memory size, we can simply return 1
-// here to count each timestamp as 1 item.
-func (c cacheableTimestamp) Size() (uint64, error) {
-	return 1, nil
-}
-
-// LndRpcChainBridge is an implementation of the tapgarden.ChainBridge
-// interface backed by an active remote lnd node.
-type LndRpcChainBridge struct {
-	lnd *lndclient.LndServices
-
-	blockTimestampCache *lru.Cache[uint32, cacheableTimestamp]
-	retryConfig         fn.RetryConfig
-
-	assetStore *tapdb.AssetStore
-}
-
-// NewLndRpcChainBridge creates a new chain bridge from an active lnd services
-// client.
-func NewLndRpcChainBridge(lnd *lndclient.LndServices,
-	assetStore *tapdb.AssetStore) *LndRpcChainBridge {
-
-	return &LndRpcChainBridge{
-		lnd: lnd,
-		blockTimestampCache: lru.NewCache[uint32, cacheableTimestamp](
-			maxNumBlocksInCache,
-		),
-		retryConfig: fn.DefaultRetryConfig(),
-		assetStore:  assetStore,
-	}
-}
-
-// RegisterConfirmationsNtfn registers an intent to be notified once
-// txid reaches numConfs confirmations.
-func (l *LndRpcChainBridge) RegisterConfirmationsNtfn(ctx context.Context,
-	txid *chainhash.Hash, pkScript []byte, numConfs, heightHint uint32,
-	includeBlock bool,
-	reOrgChan chan struct{}) (*chainntnfs.ConfirmationEvent, chan error,
-	error) {
-
-	opts := []lndclient.NotifierOption{
-		lndclient.WithReOrgChan(reOrgChan),
-	}
-	if includeBlock {
-		opts = append(opts, lndclient.WithIncludeBlock())
-	}
-
-	ctx, cancel := context.WithCancel(ctx) // nolint:govet
-	confChan, errChan, err := l.lnd.ChainNotifier.RegisterConfirmationsNtfn(
-		ctx, txid, pkScript, int32(numConfs), int32(heightHint),
-		opts...,
-	)
-	if err != nil {
-		cancel()
-
-		return nil, nil, fmt.Errorf("unable to register for conf: %w",
-			err)
-	}
-
-	return &chainntnfs.ConfirmationEvent{
-		Confirmed: confChan,
-		Cancel:    cancel,
-	}, errChan, nil
-}
-
-// RegisterBlockEpochNtfn registers an intent to be notified of each new block
-// connected to the main chain.
-func (l *LndRpcChainBridge) RegisterBlockEpochNtfn(
-	ctx context.Context) (chan int32, chan error, error) {
-
-	return l.lnd.ChainNotifier.RegisterBlockEpochNtfn(ctx)
-}
-
-// GetBlock returns a chain block given its hash.
-func (l *LndRpcChainBridge) GetBlock(ctx context.Context,
-	hash chainhash.Hash) (*wire.MsgBlock, error) {
-
-	return fn.RetryFuncN(
-		ctx, l.retryConfig, func() (*wire.MsgBlock, error) {
-			block, err := l.lnd.ChainKit.GetBlock(ctx, hash)
-			if err != nil {
-				return nil, fmt.Errorf(
-					"unable to retrieve block: %w", err,
-				)
-			}
-			return block, nil
-		},
-	)
-}
-
-// GetBlockHeader returns a block header given its hash.
-func (l *LndRpcChainBridge) GetBlockHeader(ctx context.Context,
-	hash chainhash.Hash) (*wire.BlockHeader, error) {
-
-	return fn.RetryFuncN(
-		ctx, l.retryConfig, func() (*wire.BlockHeader, error) {
-			header, err := l.lnd.ChainKit.GetBlockHeader(ctx, hash)
-			if err != nil {
-				return nil, fmt.Errorf(
-					"unable to retrieve block "+
-						"header: %w", err,
-				)
-			}
-			return header, nil
-		},
-	)
-}
-
-// GetBlockHash returns the hash of the block in the best blockchain at the
-// given height.
-func (l *LndRpcChainBridge) GetBlockHash(ctx context.Context,
-	blockHeight int64) (chainhash.Hash, error) {
-
-	return fn.RetryFuncN(
-		ctx, l.retryConfig, func() (chainhash.Hash, error) {
-			blockHash, err := l.lnd.ChainKit.GetBlockHash(
-				ctx, blockHeight,
-			)
-			if err != nil {
-				return chainhash.Hash{}, fmt.Errorf(
-					"unable to retrieve block hash: %w",
-					err,
-				)
-			}
-			return blockHash, nil
-		},
-	)
-}
-
-// VerifyBlock returns an error if a block (with given header and height) is not
-// present on-chain. It also checks to ensure that block height corresponds to
-// the given block header.
-func (l *LndRpcChainBridge) VerifyBlock(ctx context.Context,
-	header wire.BlockHeader, height uint32) error {
-
-	// TODO(ffranr): Once we've released 0.3.0, every proof should have an
-	// assigned height. At that point, we should return an error for proofs
-	// with unset (zero) block heights.
-	if height == 0 {
-		_, err := l.GetBlock(ctx, header.BlockHash())
-		return err
-	}
-
-	// Ensure that the block hash matches the hash of the block
-	// found at the given height.
-	hash, err := l.GetBlockHash(ctx, int64(height))
-	if err != nil {
-		return err
-	}
-
-	expectedHash := header.BlockHash()
-	if hash != expectedHash {
-		return fmt.Errorf("block hash and block height "+
-			"mismatch; (height: %d, hashAtHeight: %s, "+
-			"expectedHash: %s)", height, hash, expectedHash)
-	}
-
-	// Ensure that the block header corresponds to a block on-chain. Fetch
-	// only the corresponding block header and not the entire block if
-	// supported.
-	_, err = l.GetBlockHeader(ctx, header.BlockHash())
-	return err
-}
-
-// CurrentHeight return the current height of the main chain.
-func (l *LndRpcChainBridge) CurrentHeight(ctx context.Context) (uint32, error) {
-	return fn.RetryFuncN(
-		ctx, l.retryConfig, func() (uint32, error) {
-			_, bestHeight, err := l.lnd.ChainKit.GetBestBlock(ctx)
-			if err != nil {
-				return 0, fmt.Errorf(
-					"unable to grab block height: %w", err,
-				)
-			}
-			return uint32(bestHeight), nil
-		},
-	)
-}
-
-// GetBlockTimestamp returns the timestamp of the block at the given height.
-func (l *LndRpcChainBridge) GetBlockTimestamp(ctx context.Context,
-	height uint32) int64 {
-
-	// Shortcut any lookup in case we don't have a valid height in the first
-	// place.
-	if height == 0 {
-		return 0
-	}
-
-	cacheTS, err := l.blockTimestampCache.Get(height)
-	if err == nil {
-		return int64(cacheTS)
-	}
-
-	hash, err := fn.RetryFuncN(
-		ctx, l.retryConfig, func() (chainhash.Hash, error) {
-			return l.lnd.ChainKit.GetBlockHash(ctx, int64(height))
-		},
-	)
-	if err != nil {
-		return 0
-	}
-
-	// Get block header.
-	header, err := l.GetBlockHeader(ctx, hash)
-	if err != nil {
-		return 0
-	}
-
-	ts := uint32(header.Timestamp.Unix())
-	_, _ = l.blockTimestampCache.Put(height, cacheableTimestamp(ts))
-
-	return int64(ts)
-}
-
-// PublishTransaction attempts to publish a new transaction to the
-// network.
-func (l *LndRpcChainBridge) PublishTransaction(ctx context.Context,
-	tx *wire.MsgTx, label string) error {
-
-	_, err := fn.RetryFuncN(
-		ctx, l.retryConfig, func() (struct{}, error) {
-			return struct{}{}, l.lnd.WalletKit.PublishTransaction(
-				ctx, tx, label,
-			)
-		},
-	)
-	return err
-}
-
-// EstimateFee returns a fee estimate for the confirmation target.
-func (l *LndRpcChainBridge) EstimateFee(ctx context.Context,
-	confTarget uint32) (chainfee.SatPerKWeight, error) {
-
-	return fn.RetryFuncN(
-		ctx, l.retryConfig, func() (chainfee.SatPerKWeight, error) {
-			return l.lnd.WalletKit.EstimateFeeRate(
-				ctx, int32(confTarget),
-			)
-		},
-	)
-}
-
-// GenFileChainLookup generates a chain lookup interface for the given
-// proof file that can be used to validate proofs.
-func (l *LndRpcChainBridge) GenFileChainLookup(
-	f *proof.File) asset.ChainLookup {
-
-	return NewProofChainLookup(l, l.assetStore, f)
-}
-
-// GenProofChainLookup generates a chain lookup interface for the given
-// single proof that can be used to validate proofs.
-func (l *LndRpcChainBridge) GenProofChainLookup(
-	p *proof.Proof) (asset.ChainLookup, error) {
-
-	f, err := proof.NewFile(proof.V0, *p)
-	if err != nil {
-		return nil, err
-	}
-
-	return NewProofChainLookup(l, l.assetStore, f), nil
-}
-
-// A compile time assertion to ensure LndRpcChainBridge meets the
-// tapgarden.ChainBridge interface.
-var _ tapgarden.ChainBridge = (*LndRpcChainBridge)(nil)
-
-// LndMsgTransportClient is an LND RPC message transport client.
-type LndMsgTransportClient struct {
-	lnd *lndclient.LndServices
-}
-
-// NewLndMsgTransportClient creates a new message transport RPC client for a
-// given LND service.
-func NewLndMsgTransportClient(
-	lnd *lndclient.LndServices) *LndMsgTransportClient {
-
-	return &LndMsgTransportClient{
-		lnd: lnd,
-	}
-}
-
-// SubscribeCustomMessages creates a subscription to custom messages received
-// from our peers.
-func (l *LndMsgTransportClient) SubscribeCustomMessages(
-	ctx context.Context) (<-chan lndclient.CustomMessage,
-	<-chan error, error) {
-
-	return l.lnd.Client.SubscribeCustomMessages(ctx)
-}
-
-// SendCustomMessage sends a custom message to a peer.
-func (l *LndMsgTransportClient) SendCustomMessage(ctx context.Context,
-	msg lndclient.CustomMessage) error {
-
-	return l.lnd.Client.SendCustomMessage(ctx, msg)
-}
-
-// SendMessage sends a message to a remote peer.
-func (l *LndMsgTransportClient) SendMessage(ctx context.Context,
-	peer btcec.PublicKey, msg lnwire.Message) error {
-
-	var buf bytes.Buffer
-	if err := msg.Encode(&buf, 0); err != nil {
-		return fmt.Errorf("unable to encode message: %w", err)
-	}
-
-	return l.SendCustomMessage(ctx, lndclient.CustomMessage{
-		Peer:    route.NewVertex(&peer),
-		MsgType: uint32(msg.MsgType()),
-		Data:    buf.Bytes(),
-	})
-}
-
-// ReportError sends a custom message with the error type to a peer.
-//
-// NOTE: In order for this custom message to be sent over the lnd RPC interface,
-// lnd needs to be configured with the `--custom-message=17` flag, which allows
-// sending the non-custom error message type.
-func (l *LndMsgTransportClient) ReportError(ctx context.Context,
-	peer btcec.PublicKey, pid funding.PendingChanID, err error) {
-
-	srvrLog.Errorf("Error in funding flow for pending chan ID %x: %v",
-		pid[:], err)
-
-	msg := &lnwire.Error{
-		ChanID: pid,
-		Data:   []byte(err.Error()),
-	}
-
-	sendErr := l.SendMessage(ctx, peer, msg)
-	if sendErr != nil {
-		srvrLog.Errorf("Error sending error message to peer %x: %v",
-			peer.SerializeCompressed(), sendErr)
-	}
-}
-
-// Ensure LndMsgTransportClient implements the rfq.PeerMessenger,
-// tapchannel.PeerMessenger and tapchannel.ErrorReporter interfaces.
-var _ rfq.PeerMessenger = (*LndMsgTransportClient)(nil)
-var _ tapchannel.PeerMessenger = (*LndMsgTransportClient)(nil)
-var _ tapchannel.ErrorReporter = (*LndMsgTransportClient)(nil)
 
 // LndRouterClient is an LND router RPC client.
 type LndRouterClient struct {
@@ -465,184 +86,3 @@ func (l *LndInvoicesClient) HtlcModifier(ctx context.Context,
 // Ensure LndInvoicesClient implements the tapchannel.InvoiceHtlcModifier
 // interface.
 var _ tapchannel.InvoiceHtlcModifier = (*LndInvoicesClient)(nil)
-
-// LndFsmDaemonAdapters is a struct that implements the protofsm.DaemonAdapters
-// interface.
-type LndFsmDaemonAdapters struct {
-	// lnd is the LND services client that will be used to interact with
-	// the LND node.
-	lnd *lndclient.LndServices
-
-	// retryConfig is the retry configuration that will be used for
-	// operations that may fail due to temporary issues, such as network
-	// errors or RPC timeouts.
-	retryConfig fn.RetryConfig
-
-	// msgTransport is the message transport client that will be used to
-	// send messages to peers.
-	msgTransport LndMsgTransportClient
-
-	// chainBridge is the chain bridge that will be used to interact with
-	// the blockchain.
-	chainBridge LndRpcChainBridge
-
-	// ContextGuard manages the context and quit channel for this service.
-	fn.ContextGuard
-
-	startOnce sync.Once
-	stopOnce  sync.Once
-}
-
-// NewLndFsmDaemonAdapters creates a new instance of LndFsmDaemonAdapters.
-func NewLndFsmDaemonAdapters(lnd *lndclient.LndServices) *LndFsmDaemonAdapters {
-	retryConfig := fn.DefaultRetryConfig()
-
-	msgTransport := NewLndMsgTransportClient(lnd)
-
-	// Initialize the chain bridge without the asset store, as it is not
-	// needed for the FSM adapters.
-	chainBridge := NewLndRpcChainBridge(lnd, nil)
-	chainBridge.retryConfig = retryConfig
-
-	return &LndFsmDaemonAdapters{
-		lnd:          lnd,
-		retryConfig:  retryConfig,
-		msgTransport: *msgTransport,
-		chainBridge:  *chainBridge,
-		ContextGuard: fn.ContextGuard{
-			DefaultTimeout: DefaultTimeout,
-			Quit:           make(chan struct{}),
-		},
-	}
-}
-
-// Start attempts to start the service.
-func (l *LndFsmDaemonAdapters) Start() error {
-	l.startOnce.Do(func() {})
-	return nil
-}
-
-// Stop signals for the service to gracefully exit.
-func (l *LndFsmDaemonAdapters) Stop() error {
-	l.stopOnce.Do(func() {
-		// Signal the quit channel which will cancel all active
-		// contexts.
-		close(l.Quit)
-	})
-
-	return nil
-}
-
-// SendMessages sends a slice of lnwire.Message to the peer with the given
-// public key.
-func (l *LndFsmDaemonAdapters) SendMessages(peer btcec.PublicKey,
-	messages []lnwire.Message) error {
-
-	// Convert messages to a slice of CustomMessage.
-	customMessages := make([]lndclient.CustomMessage, 0, len(messages))
-	for idx := range messages {
-		msg := messages[idx]
-
-		var buf bytes.Buffer
-		if err := msg.Encode(&buf, 0); err != nil {
-			return fmt.Errorf("unable to encode message: %w", err)
-		}
-
-		customMsg := lndclient.CustomMessage{
-			Peer:    route.NewVertex(&peer),
-			MsgType: uint32(msg.MsgType()),
-			Data:    buf.Bytes(),
-		}
-
-		customMessages = append(customMessages, customMsg)
-	}
-
-	ctx, cancel := l.WithCtxQuitNoTimeout()
-	defer cancel()
-
-	// Send each message in turn.
-	for idx := range customMessages {
-		msg := customMessages[idx]
-
-		err := l.msgTransport.SendCustomMessage(ctx, msg)
-		if err != nil {
-			return fmt.Errorf("unable to send custom message: %w",
-				err)
-		}
-	}
-
-	return nil
-}
-
-// BroadcastTransaction attempts to broadcast a transaction to the
-// network. It uses the chain bridge to publish the transaction.
-func (l *LndFsmDaemonAdapters) BroadcastTransaction(tx *wire.MsgTx,
-	label string) error {
-
-	ctx, cancel := l.WithCtxQuitNoTimeout()
-	defer cancel()
-
-	return l.chainBridge.PublishTransaction(ctx, tx, label)
-}
-
-// RegisterConfirmationsNtfn registers an intent to be notified once the
-// transaction with the given txid reaches the specified number of
-// confirmations.
-func (l *LndFsmDaemonAdapters) RegisterConfirmationsNtfn(
-	txid *chainhash.Hash, pkScript []byte, numConfs uint32,
-	heightHint uint32,
-	optFuncs ...chainntnfs.NotifierOption) (*chainntnfs.ConfirmationEvent,
-	error) {
-
-	opts := chainntnfs.DefaultNotifierOptions()
-	for _, optFunc := range optFuncs {
-		optFunc(opts)
-	}
-
-	lndCliOpt := make([]lndclient.NotifierOption, 0, len(optFuncs))
-	if opts.IncludeBlock {
-		lndCliOpt = append(lndCliOpt, lndclient.WithIncludeBlock())
-	}
-
-	ctx, cancel := l.WithCtxQuitNoTimeout()
-	spendDetail, _, err := l.lnd.ChainNotifier.RegisterConfirmationsNtfn(
-		ctx, txid, pkScript, int32(numConfs), int32(heightHint),
-		lndCliOpt...,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("unable to register for conf: %w", err)
-	}
-
-	return &chainntnfs.ConfirmationEvent{
-		Confirmed:    spendDetail,
-		Updates:      make(chan uint32),
-		NegativeConf: make(chan int32),
-		Done:         make(chan struct{}),
-		Cancel:       cancel,
-	}, nil
-}
-
-// RegisterSpendNtfn registers an intent to be notified once the outpoint
-// is spent on-chain.
-func (l *LndFsmDaemonAdapters) RegisterSpendNtfn(outpoint *wire.OutPoint,
-	pkScript []byte, heightHint uint32) (*chainntnfs.SpendEvent, error) {
-
-	ctx, cancel := l.WithCtxQuitNoTimeout()
-	spendDetail, _, err := l.lnd.ChainNotifier.RegisterSpendNtfn(
-		ctx, outpoint, pkScript, int32(heightHint),
-	)
-	if err != nil {
-		return nil, fmt.Errorf("unable to register for spend: %w", err)
-	}
-
-	return &chainntnfs.SpendEvent{
-		Spend:  spendDetail,
-		Reorg:  make(chan struct{}, 1),
-		Done:   make(chan struct{}, 1),
-		Cancel: cancel,
-	}, nil
-}
-
-// Ensure LndFsmDaemonAdapters implements the protofsm.DaemonAdapters
-// interface.
-var _ protofsm.DaemonAdapters = (*LndFsmDaemonAdapters)(nil)
